Rename: common 타입 파일명 수정
